adds aws s3 upload and save products

+++ ecommerce-be/src/main/java/com/ecommerce/controller/merchant/model/MerchantProductCreateRequest.java
@@ -11,7 +11,7 @@ public record MerchantProductCreateRequest(
         @NotNull
         BigDecimal price,
         @NotNull
-        Integer inventory,
+        Long inventory,
         @NotNull
         ShipmentDeliveryTimes shipmentDeliveryTimes,
         @NotNull
